# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Management System API",
        "description": "",
        "contact": {
            "email": "novirachmahwati@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/1.0"
        }
    ],
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get all authors",
                "description": "Display a listing of the resource.",
                "operationId": "33b4ca0e528a8b6298d9942781a0c85a",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter results by string. Search author name.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/order_by"
                    },
                    {
                        "$ref": "#/components/parameters/order_direction"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of authors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/author"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Create a new author",
                "operationId": "d998cf74685b2e9609bcd48ab986f678",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/authorEditable"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Author created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/author"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get author by ID",
                "description": "Returns a single author",
                "operationId": "getAuthorById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the author to get",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/author"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Update an existing author",
                "description": "Returns updated author data",
                "operationId": "updateAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the author to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/authorEditable"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Data not found"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Delete an existing author",
                "description": "Deletes a author and returns no content",
                "operationId": "deleteAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the author to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            }
        },
        "/authors/{id}/books": {
            "get": {
                "tags": [
                    "Associations"
                ],
                "summary": "Retrieve books by author",
                "description": "Returns a list of books written by a specific author",
                "operationId": "getBooksByAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the author",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/book"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Author not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "authorEditable": {
                "properties": {
                    "name": {
                        "description": "The attributes that are mass assignable.",
                        "type": "string"
                    },
                    "bio": {
                        "type": "string"
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "author": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/authorEditable"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "bookEditable": {
                "properties": {
                    "title": {
                        "description": "The attributes that are mass assignable.",
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "publish_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "author_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "book": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/bookEditable"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        },
        "responses": {
...
            "422": {
                "description": "Unprocessable Entity",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "errors": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
...
            "order_by": {
                "name": "order_by",
                "in": "query",
                "description": "Field to order results by",
                "schema": {
                    "type": "string"
                }
            },
            "order_direction": {
                "name": "order_direction",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "asc",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "schema": {
                    "type": "integer",
                    "default": "10"
                }
            },
            "page": {
                "name": "page",
                "in": "query",
                "schema": {
                    "type": "integer",
                    "default": "1"
                }
            }
        }
    },
...
}